+++ roqoqo/src/operations/pragma_operations.rs
@@ -13,19 +13,19 @@
 //! Collection of roqoqo PRAGMA operations.
 //!
 
-use ndarray::{array, Array1, Array2};
+use crate::operations::{
+    InvolveQubits, InvolvedQubits, Operate, OperateMultiQubit, OperatePragma, OperatePragmaNoise,
+    OperateSingleQubit, RoqoqoError, Substitute,
+};
+use crate::Circuit;
+use nalgebra::Matrix4;
+use ndarray::{array, Array1, Array2, Array};
 use num_complex::Complex64;
 use qoqo_calculator::{Calculator, CalculatorFloat};
 #[cfg(feature = "serialize")]
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::convert::TryFrom;
-
-use crate::operations::{
-    InvolveQubits, InvolvedQubits, Operate, OperateMultiQubit, OperatePragma, OperatePragmaNoise,
-    OperateSingleQubit, RoqoqoError, Substitute,
-};
-use crate::Circuit;
 
 /// This PRAGMA Operation sets the number of measurements of the circuit.
 ///
@@ -740,8 +740,8 @@ impl OperatePragmaNoise for PragmaRandomNoise {
 
 /// The general noise PRAGMA operation.
 ///
-/// This PRAGMA Operation applies a noise term according to the given operators.
-/// The operators are represented by a 3x3 matrix:
+/// This PRAGMA operation applies a noise term according to the given rates.
+/// The rates are represented by a 3x3 matrix:
 /// $$ M = \begin{pmatrix}
 /// a & b & c \\\\
 /// d & e & f \\\\
@@ -750,10 +750,12 @@ impl OperatePragmaNoise for PragmaRandomNoise {
 /// where the coefficients correspond to the following summands
 /// expanded from the first term of the non-coherent part of the Lindblad equation:
 ///     $$ \frac{d}{dt}\rho = \sum_{i,j=0}^{2} M_{i,j} L_{i} \rho L_{j}^{\dagger} - \frac{1}{2} \{ L_{j}^{\dagger} L_i, \rho \} \\\\
-///         L_0 = \sigma^{-} \\\\
-///         L_1 = \sigma^{+} \\\\
+///         L_0 = \sigma^{+} \\\\
+///         L_1 = \sigma^{-} \\\\
 ///         L_3 = \sigma^{z}
 ///     $$
+///
+/// Applying the Pragma with a given `gate_time` corresponds to applying the full time-evolution under the Lindblad equation for `gate_time` time.
 ///
 /// # Example
 ///
@@ -762,7 +764,7 @@ impl OperatePragmaNoise for PragmaRandomNoise {
 /// use roqoqo::operations::PragmaGeneralNoise;
 /// use qoqo_calculator::CalculatorFloat;
 ///
-/// let operators: Array2<f64> = array![
+/// let rates: Array2<f64> = array![
 ///    [
 ///         1.0,
 ///         0.0,
@@ -782,8 +784,7 @@ impl OperatePragmaNoise for PragmaRandomNoise {
 /// let pragma = PragmaGeneralNoise::new(
 ///     0,
 ///     CalculatorFloat::from(0.005),
-///     CalculatorFloat::from(0.02),
-///     operators.clone(),
+///     rates.clone(),
 /// );
 /// ```
 /// That will result into $.
@@ -804,10 +805,8 @@ pub struct PragmaGeneralNoise {
     qubit: usize,
     /// The time (in seconds) the gate takes to be applied to the qubit on the (simulated) hardware
     gate_time: CalculatorFloat,
-    /// The error rate of the noise (in 1/second).
-    rate: CalculatorFloat,
-    /// The operators representing the general noise (a 3x3 matrix).
-    operators: Array2<f64>,
+    /// The rates representing the general noise matrix M (a 3x3 matrix).
+    rates: Array2<f64>,
 }
 
 #[allow(non_upper_case_globals)]
@@ -817,6 +816,114 @@ const TAGS_PragmaGeneralNoise: &[&str; 4] = &[
     "PragmaOperation",
     "PragmaGeneralNoise",
 ];
+
+// Collection of superoperators that appear in the Lindblad equation for a single qubit/spin with
+// a basis of the form 0: sigma+ 1:sigma- 2: sigmaz
+const PGN_SUPEROP: [[[[f64; 4]; 4]; 3]; 3] = [
+    [
+        // sigma+ sigma+
+        [
+            [-1., 0., 0., 0.],
+            [0., -0.5, 0., 0.],
+            [0., 0., -0.5, 0.],
+            [1., 0., 0., 0.],
+        ],
+        // sigma+ sigma-
+        [
+            [0., 0., 0., 0.],
+            [0., 0., 0., 0.],
+            [0., 1., 0., 0.],
+            [0., 0., 0., 0.],
+        ],
+        // sigma+ sigmaz
+        [
+            [0., 0.5, 0., 0.],
+            [0., 0., 0., 0.],
+            [1.5, 0., 0., 0.5],
+            [0., -0.5, 0., 0.],
+        ],
+    ],
+    [
+        // sigma- sigma+
+        [
+            [0., 0., 0., 0.],
+            [0., 0., 1., 0.],
+            [0., 0., 0., 0.],
+            [0., 0., 0., 0.],
+        ],
+        // sigma- sigma-
+        [
+            [0., 0., 0., 1.],
+            [0., -0.5, 0., 0.],
+            [0., 0., -0.5, 0.],
+            [0., 0., 0., -1.],
+        ],
+        // sigma- sigmaz
+        [
+            [0., 0., 0.5, 0.],
+            [-0.5, 0., 0., -1.5],
+            [0., 0., 0., 0.],
+            [0., 0., -0.5, 0.],
+        ],
+    ],
+    [
+        // sigmaz sigma+
+        [
+            [0., 0., 0.5, 0.],
+            [1.5, 0., 0., 0.5],
+            [0., 0., 0., 0.],
+            [0., 0., -0.5, 0.],
+        ],
+        // sigmaz sigma-
+        [
+            [0., 0.5, 0., 0.],
+            [0., 0., 0., 0.],
+            [-0.5, 0., 0., -1.5],
+            [0., -0.5, 0., 0.],
+        ],
+        // sigmaz sigmaz
+        [
+            [0., 0., 0., 0.],
+            [0., -2., 0., 0.],
+            [0., 0., -2., 0.],
+            [0., 0., 0., 0.],
+        ],
+    ],
+];
+
+/// OperatePragmaNoise trait creating necessary functions for a PRAGMA noise Operation.
+impl OperatePragmaNoise for PragmaGeneralNoise {
+    fn superoperator(&self) -> Result<Array2<f64>, RoqoqoError> {
+        let gate_time: f64 = f64::try_from(self.gate_time.clone())?;
+        // Creating the superoperator that propagates the density matrix in vector form scaled by rate and time
+        let mut superop = Matrix4::<f64>::default();
+        for (i, row) in PGN_SUPEROP.iter().enumerate() {
+            for (j, op) in row.iter().clone().enumerate() {
+                let tmp_superop: Matrix4<f64> = (*op).into();
+                superop +=  gate_time * self.rates[(i, j)] * tmp_superop;
+            }
+        }
+        // Integrate superoperator for infinitesimal time to get superoperator for given rate and gate-time
+        // Use exponential
+        let exp_superop: Matrix4<f64> = superop.exp();
+        let mut tmp_iter = exp_superop.iter();
+        // convert to ndarray.
+        let array: Array2<f64> = Array::from_shape_simple_fn((4,4), || *tmp_iter.next().unwrap());
+
+        Ok(array)
+    }
+
+    fn probability(&self) -> CalculatorFloat {
+        CalculatorFloat::ZERO
+    }
+
+    /// Returns the gate to the power of `power`.
+    fn powercf(&self, power: CalculatorFloat) -> Self {
+        let mut new = self.clone();
+        new.gate_time = power * self.gate_time.clone();
+        new
+    }
+}
 
 /// The conditional PRAGMA operation.
 ///