+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This changelog track changes to the qoqo project starting at version 0.5.0
 ### Changed
 
 * Rarely used qubit mapping is now the last argument in PragmaRepeatedMeasurement
-* PragmaGeneralNoise uses sigma^+ sigma^- and sigma^z as a basis to for Lindblad decoherence rates to avoid using complex rates.
+* PragmaGeneralNoise uses sigma^+ sigma^- and sigma^z as a basis to for Lindblad decoherence rates to avoid using complex rates. Rate and operators parameters of PragmaGeneralNoise have been combined in single parameter rates.
 
 ### Added